requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
werkzeug==2.3.7
gunicorn==21.2.0